Factor out filetype-specific tests from TMServiceTest
