+++ webapp/src/test/java/com/box/l10n/mojito/service/tm/TMServiceTest.java
@@ -3,7 +3,6 @@ package com.box.l10n.mojito.service.tm;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
@@ -18,15 +17,12 @@ import com.box.l10n.mojito.entity.TMTextUnitCurrentVariant;
 import com.box.l10n.mojito.entity.TMTextUnitVariant;
 import com.box.l10n.mojito.entity.TMTextUnitVariantComment;
 import com.box.l10n.mojito.entity.TMXliff;
-import com.box.l10n.mojito.okapi.FilterConfigIdOverride;
 import com.box.l10n.mojito.okapi.ImportTranslationsFromLocalizedAssetStep.StatusForEqualTarget;
 import com.box.l10n.mojito.okapi.InheritanceMode;
 import com.box.l10n.mojito.okapi.Status;
 import com.box.l10n.mojito.okapi.TextUnitUtils;
 import com.box.l10n.mojito.okapi.XliffState;
 import com.box.l10n.mojito.okapi.asset.UnsupportedAssetFilterTypeException;
-import com.box.l10n.mojito.okapi.filters.AndroidXMLEncoder;
-import com.box.l10n.mojito.okapi.filters.SimpleEncoder;
 import com.box.l10n.mojito.service.asset.AssetRepository;
 import com.box.l10n.mojito.service.asset.AssetService;
 import com.box.l10n.mojito.service.asset.AssetUpdateException;
@@ -59,7 +55,6 @@ import net.sf.okapi.common.resource.TextUnit;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
 import org.hibernate.proxy.HibernateProxy;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.slf4j.Logger;
@@ -1418,2789 +1413,6 @@ public class TMServiceTest extends ServiceTestBase {
     return cleanXliff;
   }
 
-  /**
-   * This test is to test {@link AndroidXMLEncoder} with option to override encoding of &lt; and
-   * &gt;
-   *
-   * <p>According to Android specification in
-   * http://developer.android.com/guide/topics/resources/string-resource.html, <b>bold</b>,
-   * <i>italic</i> and <u>underline</u> should be in localized file as-is.
-   *
-   * @throws Exception
-   */
-  @Test
-  public void testLocalizeAndroidStringsWithSpecialCharactersOldEscaping() throws Exception {
-
-    List<String> filterOptionOldEscaping = Arrays.asList("oldEscaping=true");
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "    <string description=\"Example html markup string1\" name=\"welcome1\">Welcome to <b>Android</b>!</string>\n"
-            + "    <string description=\"Example html markup string2\" name=\"welcome2\">Welcome to <i>Android</i>!</string>\n"
-            + "    <string description=\"Example html markup string3\" name=\"welcome3\">Welcome to <u>Android</u>!</string>\n"
-            + "    <string description=\"Example html markup string4\" name=\"welcome4\">Welcome to <annotation font=\"title_emphasis\">Android</annotation>!</string>\n"
-            + "    <string name=\"subheader_text1\">\\\'Make sure you\\\'d \\\"escaped\\\" special characters like quotes &amp; ampersands.\\n</string>\n"
-            + "    <string name=\"subheader_text2\">\"This'll also work\"</string>\n"
-            + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(),
-            asset.getPath(),
-            assetContent,
-            false,
-            null,
-            null,
-            null,
-            null,
-            null,
-            filterOptionOldEscaping);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            filterOptionOldEscaping,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeAndroidStringsWithSpecialCharacters() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "    <string description=\"Text for when user is drawing.\" name=\"test_annotation\"><annotation tag=\"user\">%1$s</annotation> is drawing…</string>"
-            + "    <string description=\"Example html markup string1\" name=\"welcome1\">Welcome to <b>Android</b>!</string>\n"
-            + "    <string description=\"Example html markup string2\" name=\"welcome2\">Welcome to <i>Android</i>!</string>\n"
-            + "    <string description=\"Example html markup string3\" name=\"welcome3\">Welcome to <u>Android</u>!</string>\n"
-            + "    <string description=\"Example html markup string4\" name=\"welcome4\">Welcome to <annotation font=\"title_emphasis\">Android</annotation>!</string>\n"
-            + "    <string name=\"subheader_text1\">\\\'Make sure you\\\'d \\\"escaped\\\" special characters like quotes &amp; ampersands.\\n</string>\n"
-            + "    <string name=\"subheader_text2\">\"This'll also work\"</string>\n"
-            + "    <string name=\"escape_dot\">\\.</string>\n"
-            + "    <string name=\"escape_quote\">\\'</string>\n"
-            + "    <string name=\"escape_apostrophe\">\\ʼ</string>\n"
-            + "    <string name=\"escape_escape\">\\\\</string>\n"
-            + "    <string name=\"escape_at_sign\">\\@</string>\n"
-            + "    <string name=\"escape_ampersand\">&amp;</string>\n"
-            + "    <string name=\"escape_lowerthan\">&lt;</string>\n"
-            + "    <string name=\"replace_tab\">a\tb\tc</string>\n"
-            + "    <string name=\"remove_line_feed\">\nline1\n   line2\n  line3   </string>\n"
-            + "    <string name=\"escape_line_feed2\">\\nline1\\n   line2\\n  line3   </string>\n"
-            + "    <string name=\"trim\">    \n a \n   </string>\n"
-            + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-
-    String expected =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "    <string description=\"Text for when user is drawing.\" name=\"test_annotation\"><annotation tag=\"user\">%1$s</annotation> is drawing…</string>"
-            + "    <string description=\"Example html markup string1\" name=\"welcome1\">Welcome to <b>Android</b>!</string>\n"
-            + "    <string description=\"Example html markup string2\" name=\"welcome2\">Welcome to <i>Android</i>!</string>\n"
-            + "    <string description=\"Example html markup string3\" name=\"welcome3\">Welcome to <u>Android</u>!</string>\n"
-            + "    <string description=\"Example html markup string4\" name=\"welcome4\">Welcome to <annotation font=\"title_emphasis\">Android</annotation>!</string>\n"
-            + "    <string name=\"subheader_text1\">\\'Make sure you\\'d \\\"escaped\\\" special characters like quotes &amp; ampersands.\\n</string>\n"
-            + "    <string name=\"subheader_text2\">This\\'ll also work</string>\n"
-            + "    <string name=\"escape_dot\">.</string>\n"
-            + "    <string name=\"escape_quote\">\\'</string>\n"
-            + "    <string name=\"escape_apostrophe\">ʼ</string>\n"
-            + "    <string name=\"escape_escape\">\\</string>\n"
-            + "    <string name=\"escape_at_sign\">@</string>\n"
-            + "    <string name=\"escape_ampersand\">&amp;</string>\n"
-            + "    <string name=\"escape_lowerthan\">&lt;</string>\n"
-            + "    <string name=\"replace_tab\">a b c</string>\n"
-            + "    <string name=\"remove_line_feed\">line1 line2 line3</string>\n"
-            + "    <string name=\"escape_line_feed2\">\\nline1\\n line2\\n line3</string>\n"
-            + "    <string name=\"trim\">a</string>\n"
-            + "</resources>";
-
-    assertEquals(expected, localizedAsset);
-  }
-
-  /**
-   * This test is to test {@link AndroidXMLEncoder} with escaped HTML and CDATA
-   *
-   * @throws Exception
-   */
-  @Test
-  public void testLocalizeAndroidStringsWithEscapedHTMLAndCDATA() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "    <string name=\"welcome_messages0\">Hello, %1$s! You have <b>%2$d new messages</b>.</string>\n"
-            + "    <string name=\"welcome_messages1\">Hello, %1$s! You have &lt;b>%2$d new messages&lt;/b>.</string>\n"
-            + "    <string name=\"welcome_messages2\">Hello, %1$s! You have <![CDATA[<b>%2$d new messages</b>]]>.</string>\n"
-            + "</resources>";
-    String expectedLocalizedAsset =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "    <string name=\"welcome_messages0\">Hello, %1$s! You have &lt;b>%2$d new messages&lt;/b>.</string>\n"
-            + "    <string name=\"welcome_messages1\">Hello, %1$s! You have &lt;b>%2$d new messages&lt;/b>.</string>\n"
-            + "    <string name=\"welcome_messages2\">Hello, %1$s! You have &lt;b>%2$d new messages&lt;/b>.</string>\n"
-            + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-      assertEquals("Hello, %1$s! You have <b>%2$d new messages</b>.", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeAndroidCommentWithTranslatableFalse() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "    <!-- Comment that should be skipped -->\n"
-            + "    <string name=\"to_skip\" translatable=\"false\">Some string to skip</string>\n"
-            + "    <!-- Comment for hello string -->\n"
-            + "    <string name=\"hello\">Hello</string>\n"
-            + "</resources>";
-
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("comment=[{}]", textUnitDTO.getComment());
-    }
-    assertEquals(1, textUnitDTOs.size());
-    assertEquals("Hello", textUnitDTOs.get(0).getSource());
-    assertEquals("Comment for hello string", textUnitDTOs.get(0).getComment());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeAndroidTranslatableFalse() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        """
-        <?xml version="1.0" encoding="UTF-8"?>
-        <resources>
-            <string name="do_not_translate" translatable="false">do_not_translate</string>
-
-            <plurals name="plural_do_not_translate" translatable="false">
-                <item formatted="false" quantity="one">plural_do_not_translate_one</item>
-                <item formatted="false" quantity="other">plural_do_not_translate_other</item>
-            </plurals>
-        </resources>""";
-
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.info("name=[{}]", textUnitDTO.getName());
-    }
-    assertEquals(2, textUnitDTOs.size());
-    // TODO translatable on plural strings don't work
-
-  }
-
-  @Test
-  public void testLocalizeAndroidUnicodeEscape() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        """
-        <?xml version="1.0" encoding="UTF-8"?>
-        <resources>
-            <!-- Test Unicode Escapes -->
-            <string name="unicode_escape">A string with\\u00A0Unicode Escape</string>
-            <string name="unicode_escape2">A string with&#x00a0;Unicode Escape</string>
-            <string name="unicode_escape3">A string with&#160;Unicode Escape</string>
-            <string name="unicode_escape4">A string with&#xa0;Unicode Escape</string>
-        </resources>""";
-
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("comment=[{}]", textUnitDTO.getComment());
-    }
-    assertEquals(4, textUnitDTOs.size());
-    assertEquals("A string with\u00A0Unicode Escape", textUnitDTOs.get(0).getSource());
-    assertEquals("Test Unicode Escapes", textUnitDTOs.get(0).getComment());
-    assertEquals("A string with\u00A0Unicode Escape", textUnitDTOs.get(1).getSource());
-    assertEquals("A string with\u00A0Unicode Escape", textUnitDTOs.get(2).getSource());
-    assertEquals("A string with\u00A0Unicode Escape", textUnitDTOs.get(3).getSource());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.error("localized=\n{}", localizedAsset);
-    String expectedLocalizedAsset =
-        """
-        <?xml version="1.0" encoding="UTF-8"?>
-        <resources>
-            <!-- Test Unicode Escapes -->
-            <string name="unicode_escape">A string with&#x00a0;Unicode Escape</string>
-            <string name="unicode_escape2">A string with&#x00a0;Unicode Escape</string>
-            <string name="unicode_escape3">A string with&#x00a0;Unicode Escape</string>
-            <string name="unicode_escape4">A string with&#x00a0;Unicode Escape</string>
-        </resources>""";
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeAndroidStringsPlural() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "ja-JP");
-      repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "  <!-- Example of plurals -->\n"
-            + "  <plurals name=\"numberOfCollaborators\">\n"
-            + "    <item quantity=\"one\">%1$d people</item>\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "  <!-- Example2 of plurals -->\n"
-            + "  <plurals name=\"numberOfCollaborators2\">\n"
-            + "    <item quantity=\"one\">%1$d people</item>\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "  <plurals name=\"numberOfCollaborators3\">\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "</resources>";
-    String expectedLocalizedAsset_jaJP =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "  <!-- Example of plurals -->\n"
-            + "  <plurals name=\"numberOfCollaborators\">\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "  <!-- Example2 of plurals -->\n"
-            + "  <plurals name=\"numberOfCollaborators2\">\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "  <plurals name=\"numberOfCollaborators3\">\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "</resources>";
-    String expectedLocalizedAsset_enGB =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "  <!-- Example of plurals -->\n"
-            + "  <plurals name=\"numberOfCollaborators\">\n"
-            + "    <item quantity=\"one\">%1$d people</item>\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "  <!-- Example2 of plurals -->\n"
-            + "  <plurals name=\"numberOfCollaborators2\">\n"
-            + "    <item quantity=\"one\">%1$d people</item>\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "  <plurals name=\"numberOfCollaborators3\">\n"
-            + "    <item quantity=\"one\">%1$d people</item>\n"
-            + "    <item quantity=\"other\">%1$d people</item>\n"
-            + "  </plurals>\n"
-            + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source [{}]=[{}]", textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset_jaJP, localizedAsset);
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset_enGB, localizedAsset);
-  }
-
-  /**
-   * This test is to test AndroidStrings array with empty item
-   *
-   * @throws Exception
-   */
-  @Test
-  public void testLocalizeAndroidStringsArrayWithEmptyItem() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<resources>\n"
-            + "    <string-array name=\"N_items_failed_to_move\">\n"
-            + "        <item/>\n"
-            + "        <item>1 item failed to move</item>\n"
-            + "        <item>%1$d items failed to move</item>\n"
-            + "    </string-array>\n"
-            + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  /**
-   * This test is to test AndroidStrings array with no xml version
-   *
-   * @throws Exception
-   */
-  @Test
-  public void testLocalizeAndroidStringsNoXMLVersion() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<resources>\n" + "    <string name=\"test\">This is test</string>\n" + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  /**
-   * This test is to test AndroidStrings with REMOVE_UNTRANSLATED inheritance mode
-   *
-   * @throws Exception
-   */
-  @Test
-  public void testLocalizeAndroidStringsRemoveUntranslatedOldEsaping() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<resources>\n"
-            + "    <string name=\"test\">This is test</string>\n"
-            + "    <string name=\"desc\">This is a description</string>\n"
-            + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    String forImport =
-        "<resources>\n" + "    <string name=\"test\">Le test</string>\n" + "</resources>\n";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            List.of("postProcessIndent=4"),
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-  }
-
-  /**
-   * This test is to test AndroidStrings with REMOVE_UNTRANSLATED inheritance mode with a single
-   * item We need a special case in {@link com.box.l10n.mojito.okapi.TranslateStep} to keep the part
-   * of the skeleton that contains the begining of the document when skipping the text unit.
-   *
-   * @throws Exception
-   */
-  @Test
-  public void testLocalizeAndroidStringsRemoveUntranslatedSingleItem() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<!-- comment after prolog -->\n\n"
-            + "<resources>\n"
-            + "    <string name=\"test\">This is test</string>\n"
-            + "</resources>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "res/values/strings.xml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    String expectedLocalized =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<!-- comment after prolog -->"
-            + "<resources/>\n";
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalized, localizedAsset);
-  }
-
-  /**
-   * This test is to test {@link SimpleEncoder} with special characters
-   *
-   * <p>According to iOS specification in
-   * https://developer.apple.com/library/ios/documentation/Cocoa/Conceptual/LoadingResources/Strings/Strings.html,
-   * the following characters should be escaped with backslash: double-quote, backslash,
-   * newline(\n), carriage return (\r).
-   *
-   * @throws Exception
-   */
-  @Test
-  public void testLocalizeMacStringsWithSpecialCharacters() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "\"100_character_description\" = \"\\\"100\\\" character description:\";\n"
-            + "\"two_lines\" = \"first\\nsecond\";";
-    asset =
-        assetService.createAssetWithContent(
-            repo.getId(), "en.lproj/Localizable.strings", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  @Test
-  public void testMacStringsdict() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale_ja;
-    RepositoryLocale repoLocale_en;
-
-    try {
-      repoLocale_ja = repositoryService.addRepositoryLocale(repo, "ja-JP");
-      repoLocale_en = repositoryService.addRepositoryLocale(repo, "en-GB");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>plural_recipe_cook_hours</key>\n"
-            + "<dict>\n"
-            + "    <key>NSStringLocalizedFormatKey</key>\n"
-            + "    <string>%#@hours@ to cook</string>\n"
-            + "    <key>hours</key>\n"
-            + "    <dict>\n"
-            + "        <key>NSStringFormatSpecTypeKey</key>\n"
-            + "        <string>NSStringPluralRuleType</string>\n"
-            + "        <key>NSStringFormatValueTypeKey</key>\n"
-            + "        <string>d</string>\n"
-            + "        <key>one</key>\n"
-            + "        <string>%d hour to cook</string>\n"
-            + "        <key>other</key>\n"
-            + "        <string>%d hours to cook</string>\n"
-            + "    </dict>\n"
-            + "</dict>\n"
-            + "<key>collaborators</key>\n"
-            + "<dict>\n"
-            + "    <key>NSStringLocalizedFormatKey</key>\n"
-            + "    <string>%#@collaborators@</string>\n"
-            + "    <key>collaborators</key>\n"
-            + "    <dict>\n"
-            + "        <key>NSStringFormatSpecTypeKey</key>\n"
-            + "        <string>NSStringPluralRuleType</string>\n"
-            + "        <key>NSStringFormatValueTypeKey</key>\n"
-            + "        <string>d</string>\n"
-            + "        <key>one</key>\n"
-            + "        <string>%d collaborator</string>\n"
-            + "        <key>other</key>\n"
-            + "        <string>%d collaborators</string>\n"
-            + "    </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-    String expectedLocalizedAsset_jaJP =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>plural_recipe_cook_hours</key>\n"
-            + "<dict>\n"
-            + "    <key>NSStringLocalizedFormatKey</key>\n"
-            + "    <string>%#@hours@ to cook</string>\n"
-            + "    <key>hours</key>\n"
-            + "    <dict>\n"
-            + "        <key>NSStringFormatSpecTypeKey</key>\n"
-            + "        <string>NSStringPluralRuleType</string>\n"
-            + "        <key>NSStringFormatValueTypeKey</key>\n"
-            + "        <string>d</string>\n"
-            + "        <key>other</key>\n"
-            + "        <string>%d hours to cook</string>\n"
-            + "    </dict>\n"
-            + "</dict>\n"
-            + "<key>collaborators</key>\n"
-            + "<dict>\n"
-            + "    <key>NSStringLocalizedFormatKey</key>\n"
-            + "    <string>%#@collaborators@</string>\n"
-            + "    <key>collaborators</key>\n"
-            + "    <dict>\n"
-            + "        <key>NSStringFormatSpecTypeKey</key>\n"
-            + "        <string>NSStringPluralRuleType</string>\n"
-            + "        <key>NSStringFormatValueTypeKey</key>\n"
-            + "        <string>d</string>\n"
-            + "        <key>other</key>\n"
-            + "        <string>%d collaborators</string>\n"
-            + "    </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-    String expectedLocalizedAsset_enGB =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>plural_recipe_cook_hours</key>\n"
-            + "<dict>\n"
-            + "    <key>NSStringLocalizedFormatKey</key>\n"
-            + "    <string>%#@hours@ to cook</string>\n"
-            + "    <key>hours</key>\n"
-            + "    <dict>\n"
-            + "        <key>NSStringFormatSpecTypeKey</key>\n"
-            + "        <string>NSStringPluralRuleType</string>\n"
-            + "        <key>NSStringFormatValueTypeKey</key>\n"
-            + "        <string>d</string>\n"
-            + "        <key>one</key>\n"
-            + "        <string>%d hour to cook</string>\n"
-            + "        <key>other</key>\n"
-            + "        <string>%d hours to cook</string>\n"
-            + "    </dict>\n"
-            + "</dict>\n"
-            + "<key>collaborators</key>\n"
-            + "<dict>\n"
-            + "    <key>NSStringLocalizedFormatKey</key>\n"
-            + "    <string>%#@collaborators@</string>\n"
-            + "    <key>collaborators</key>\n"
-            + "    <dict>\n"
-            + "        <key>NSStringFormatSpecTypeKey</key>\n"
-            + "        <string>NSStringPluralRuleType</string>\n"
-            + "        <key>NSStringFormatValueTypeKey</key>\n"
-            + "        <string>d</string>\n"
-            + "        <key>one</key>\n"
-            + "        <string>%d collaborator</string>\n"
-            + "        <key>other</key>\n"
-            + "        <string>%d collaborators</string>\n"
-            + "    </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "Localizable.stringsdict", assetContent);
-
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source [{}]=[{}]", textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale_ja,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset_jaJP, localizedAsset);
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale_en,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset_enGB, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeMacStringsdictPluralJp() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    String bcp47Tag = "ja-JP";
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, bcp47Tag);
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>%d file(s) remaining</key>\n"
-            + "<dict>\n"
-            + "   <key>NSStringLocalizedFormatKey</key>\n"
-            + "   <string>%#@files@</string>\n"
-            + "   <key>files</key>\n"
-            + "   <dict>\n"
-            + "       <key>NSStringFormatSpecTypeKey</key>\n"
-            + "       <string>NSStringPluralRuleType</string>\n"
-            + "       <key>NSStringFormatValueTypeKey</key>\n"
-            + "       <string>d</string>\n"
-            + "       <key>one</key>\n"
-            + "       <string>%d file remaining</string>\n"
-            + "       <key>other</key>\n"
-            + "       <string>%d files remaining</string>\n"
-            + "   </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-
-    String expectedLocalizedAsset =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>%d file(s) remaining</key>\n"
-            + "<dict>\n"
-            + "   <key>NSStringLocalizedFormatKey</key>\n"
-            + "   <string>%#@files@</string>\n"
-            + "   <key>files</key>\n"
-            + "   <dict>\n"
-            + "       <key>NSStringFormatSpecTypeKey</key>\n"
-            + "       <string>NSStringPluralRuleType</string>\n"
-            + "       <key>NSStringFormatValueTypeKey</key>\n"
-            + "       <string>d</string>\n"
-            + "       <key>other</key>\n"
-            + "       <string>%d files remaining</string>\n"
-            + "   </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "Localizable.stringsdict", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            bcp47Tag,
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>%d file(s) remaining</key>\n"
-            + "<dict>\n"
-            + "   <key>NSStringLocalizedFormatKey</key>\n"
-            + "   <string>%#@files@</string>\n"
-            + "   <key>files</key>\n"
-            + "   <dict>\n"
-            + "       <key>NSStringFormatSpecTypeKey</key>\n"
-            + "       <string>NSStringPluralRuleType</string>\n"
-            + "       <key>NSStringFormatValueTypeKey</key>\n"
-            + "       <string>d</string>\n"
-            + "       <key>other</key>\n"
-            + "       <string>%d files remaining-jp</string>\n"
-            + "   </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            bcp47Tag,
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeMacStringsdictPluralWithDifferentIdentation()
-      throws RepositoryNameAlreadyUsedException,
-          UnsupportedAssetFilterTypeException,
-          ExecutionException,
-          InterruptedException {
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    String bcp47Tag = "fr-FR";
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, bcp47Tag);
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<!DOCTYPE plist PUBLIC \"-//Apple//DTD PLIST 1.0//EN\" \"http://www.apple.com/DTDs/PropertyList-1.0.dtd\">\n"
-            + "<plist version=\"1.0\">\n"
-            + "    <dict>\n"
-            + "        <key>%lld follower(s)</key>\n"
-            + "        <dict>\n"
-            + "            <key>NSStringLocalizedFormatKey</key>\n"
-            + "            <string>%#@followers@</string>\n"
-            + "            <key>followers</key>\n"
-            + "            <dict>\n"
-            + "                <key>NSStringFormatSpecTypeKey</key>\n"
-            + "                <string>NSStringPluralRuleType</string>\n"
-            + "                <key>NSStringFormatValueTypeKey</key>\n"
-            + "                <string>lld</string>\n"
-            + "                <key>one</key>\n"
-            + "                <string>%lld follower</string>\n"
-            + "                <key>other</key>\n"
-            + "                <string>%lld followers</string>\n"
-            + "            </dict>\n"
-            + "        </dict>\n"
-            + "        <key>%lld following(s)</key>\n"
-            + "        <dict>\n"
-            + "            <key>NSStringLocalizedFormatKey</key>\n"
-            + "            <string>%#@following@</string>\n"
-            + "            <key>following</key>\n"
-            + "            <dict>\n"
-            + "                <key>NSStringFormatSpecTypeKey</key>\n"
-            + "                <string>NSStringPluralRuleType</string>\n"
-            + "                <key>NSStringFormatValueTypeKey</key>\n"
-            + "                <string>lld</string>\n"
-            + "                <key>one</key>\n"
-            + "                <string>%lld following</string>\n"
-            + "                <key>other</key>\n"
-            + "                <string>%lld following</string>\n"
-            + "            </dict>\n"
-            + "        </dict>\n"
-            + "    </dict>\n"
-            + "</plist>\n";
-
-    // notice the doc type is gone
-    String expectedLocalizedAsset =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<plist version=\"1.0\">\n"
-            + "    <dict>\n"
-            + "        <key>%lld follower(s)</key>\n"
-            + "        <dict>\n"
-            + "            <key>NSStringLocalizedFormatKey</key>\n"
-            + "            <string>%#@followers@</string>\n"
-            + "            <key>followers</key>\n"
-            + "            <dict>\n"
-            + "                <key>NSStringFormatSpecTypeKey</key>\n"
-            + "                <string>NSStringPluralRuleType</string>\n"
-            + "                <key>NSStringFormatValueTypeKey</key>\n"
-            + "                <string>lld</string>\n"
-            + "                <key>one</key>\n"
-            + "                <string>%lld follower</string>\n"
-            + "                <key>other</key>\n"
-            + "                <string>%lld followers</string>\n"
-            + "            </dict>\n"
-            + "        </dict>\n"
-            + "        <key>%lld following(s)</key>\n"
-            + "        <dict>\n"
-            + "            <key>NSStringLocalizedFormatKey</key>\n"
-            + "            <string>%#@following@</string>\n"
-            + "            <key>following</key>\n"
-            + "            <dict>\n"
-            + "                <key>NSStringFormatSpecTypeKey</key>\n"
-            + "                <string>NSStringPluralRuleType</string>\n"
-            + "                <key>NSStringFormatValueTypeKey</key>\n"
-            + "                <string>lld</string>\n"
-            + "                <key>one</key>\n"
-            + "                <string>%lld following</string>\n"
-            + "                <key>other</key>\n"
-            + "                <string>%lld following</string>\n"
-            + "            </dict>\n"
-            + "        </dict>\n"
-            + "    </dict>\n"
-            + "</plist>";
-
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "Localizable.stringsdict", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source [{}]=[{}]", textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "fr-FR",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "fr-FR",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeMacStringsdictPluralRu() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    String bcp47Tag = "ru-RU";
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, bcp47Tag);
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>%d file(s) remaining</key>\n"
-            + "<dict>\n"
-            + "   <key>NSStringLocalizedFormatKey</key>\n"
-            + "   <string>%#@files@</string>\n"
-            + "   <key>files</key>\n"
-            + "   <dict>\n"
-            + "       <key>NSStringFormatSpecTypeKey</key>\n"
-            + "       <string>NSStringPluralRuleType</string>\n"
-            + "       <key>NSStringFormatValueTypeKey</key>\n"
-            + "       <string>d</string>\n"
-            + "       <key>one</key>\n"
-            + "       <string>%d file remaining</string>\n"
-            + "       <key>other</key>\n"
-            + "       <string>%d files remaining</string>\n"
-            + "   </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-
-    String expectedLocalizedAsset =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>%d file(s) remaining</key>\n"
-            + "<dict>\n"
-            + "   <key>NSStringLocalizedFormatKey</key>\n"
-            + "   <string>%#@files@</string>\n"
-            + "   <key>files</key>\n"
-            + "   <dict>\n"
-            + "       <key>NSStringFormatSpecTypeKey</key>\n"
-            + "       <string>NSStringPluralRuleType</string>\n"
-            + "       <key>NSStringFormatValueTypeKey</key>\n"
-            + "       <string>d</string>\n"
-            + "       <key>one</key>\n"
-            + "       <string>%d file remaining</string>\n"
-            + "       <key>few</key>\n"
-            + "       <string>%d files remaining</string>\n"
-            + "       <key>many</key>\n"
-            + "       <string>%d files remaining</string>\n"
-            + "       <key>other</key>\n"
-            + "       <string>%d files remaining</string>\n"
-            + "   </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "Localizable.stringsdict", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            bcp47Tag,
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "<plist version=\"1.0\">\n"
-            + "<dict>\n"
-            + "<key>%d file(s) remaining</key>\n"
-            + "<dict>\n"
-            + "   <key>NSStringLocalizedFormatKey</key>\n"
-            + "   <string>%#@files@</string>\n"
-            + "   <key>files</key>\n"
-            + "   <dict>\n"
-            + "       <key>NSStringFormatSpecTypeKey</key>\n"
-            + "       <string>NSStringPluralRuleType</string>\n"
-            + "       <key>NSStringFormatValueTypeKey</key>\n"
-            + "       <string>d</string>\n"
-            + "       <key>one</key>\n"
-            + "       <string>%d file remaining-ru</string>\n"
-            + "       <key>few</key>\n"
-            + "       <string>%d files remaining-ru</string>\n"
-            + "       <key>many</key>\n"
-            + "       <string>%d files remaining-ru</string>\n"
-            + "       <key>other</key>\n"
-            + "       <string>%d files remaining-ru</string>\n"
-            + "   </dict>\n"
-            + "</dict>\n"
-            + "</dict>\n"
-            + "</plist>";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            bcp47Tag,
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizePoPluralJp() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "ja-JP");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=2; plural=(n != 1);\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"\"\n"
-            + "msgstr[1] \"\"";
-
-    String expectedLocalizedAsset =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=1; plural=0;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"repins\"\n";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "messages.pot", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=1; plural=0;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"repin-jp\"\n";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizePoRemoveUntranslated() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "ja-JP");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=2; plural=(n != 1);\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n\n\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgstr \"\"\n"
-            + "#. Description\n"
-            + "#: core/logic/week_in_review_email_logic.py:50\n"
-            + "msgid \"description\"\n"
-            + "msgstr \"\"\n";
-
-    String expectedLocalizedAsset =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=1; plural=0;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "\n"
-            + "\n";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "messages.pot", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.debug("localized=\n{}\nEOL", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=1; plural=0;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n\n\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgstr \"repin-jp\"\n";
-
-    logger.debug("formimport=\n{}", forImport);
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.info("localized after import=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizePoEscaping() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "ja-JP");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=2; plural=(n != 1);\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin \\\"{}\\\"\"\n"
-            + "msgstr \"\"";
-
-    String expectedLocalizedAsset =
-        "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=1; plural=0;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin \\\"{}\\\"\"\n"
-            + "msgstr \"repin \\\"{}\\\"\"\n";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "messages.pot", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    assertEquals(1, textUnitDTOs.size());
-    TextUnitDTO textUnitDTO = textUnitDTOs.get(0);
-    assertEquals("repin \\\"{}\\\"", textUnitDTO.getName());
-    assertEquals("repin \"{}\"", textUnitDTO.getSource());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=1; plural=0;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin \\\"{}\\\"\"\n"
-            + "msgstr \"repin \\\"{}\\\" jp\"\n";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-
-    textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.TRANSLATED);
-    textUnitSearcherParameters.setLocaleId(repoLocale.getLocale().getId());
-    textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    assertEquals(1, textUnitDTOs.size());
-    textUnitDTO = textUnitDTOs.get(0);
-    assertEquals("repin \\\"{}\\\"", textUnitDTO.getName());
-    assertEquals("repin \"{}\"", textUnitDTO.getSource());
-    assertEquals("repin \"{}\" jp", textUnitDTO.getTarget());
-  }
-
-  @Test
-  public void testLocalizePoPluralRu() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "ru-RU");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=2; plural=(n != 1);\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"\"\n"
-            + "msgstr[1] \"\"";
-
-    String expectedLocalizedAsset =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=3; plural=n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"repin\"\n"
-            + "msgstr[1] \"repins\"\n"
-            + "msgstr[2] \"repins\"\n";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "messages.pot", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    //        assertEquals("Hello, %1$s! You have <b>%2$d new messages</b>.",
-    // textUnitDTO.getSource());
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ru-RU",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=3; plural=n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"repin-ru\"\n"
-            + "msgstr[1] \"repins-ru-1\"\n"
-            + "msgstr[2] \"repins-ru-2\"\n";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ru-RU",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizePoPluralCs() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "cs-CZ");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=2; plural=(n != 1);\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"\"\n"
-            + "msgstr[1] \"\"";
-
-    String expectedLocalizedAsset =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=3; plural=(n==1) ? 0 : (n>=2 && n<=4) ? 1 : 2;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"repin\"\n"
-            + "msgstr[1] \"repins\"\n"
-            + "msgstr[2] \"repins\"\n";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "messages.pot", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "cs-CZ",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=3; plural=(n==1) ? 0 : (n>=2 && n<=4) ? 1 : 2;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin\"\n"
-            + "msgid_plural \"repins\"\n"
-            + "msgstr[0] \"repin-cs\"\n"
-            + "msgstr[1] \"repins-cz-1\"\n"
-            + "msgstr[2] \"repins-cz-2\"\n";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "cs-CZ",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizePoPluralAr() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "ar-SA");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=2; plural=(n != 1);\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin \\\"{placeholder}\\\"\"\n"
-            + "msgid_plural \"repins \\\"{placeholder}\\\"\"\n"
-            + "msgstr[0] \"\"\n"
-            + "msgstr[1] \"\"";
-
-    String expectedLocalizedAsset =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=6; plural=n==0 ? 0 : n==1 ? 1 : n==2 ? 2 : n%100>=3 && n%100<=10 ? 3 : n%100>=11 ? 4 : 5;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin \\\"{placeholder}\\\"\"\n"
-            + "msgid_plural \"repins \\\"{placeholder}\\\"\"\n"
-            + "msgstr[0] \"repins \\\"{placeholder}\\\"\"\n"
-            + "msgstr[1] \"repin \\\"{placeholder}\\\"\"\n"
-            + "msgstr[2] \"repins \\\"{placeholder}\\\"\"\n"
-            + "msgstr[3] \"repins \\\"{placeholder}\\\"\"\n"
-            + "msgstr[4] \"repins \\\"{placeholder}\\\"\"\n"
-            + "msgstr[5] \"repins \\\"{placeholder}\\\"\"\n";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "messages.pot", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ar-SA",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-09-15 11:53-0500\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=6; plural=n==0 ? 0 : n==1 ? 1 : n==2 ? 2 : n%100>=3 && n%100<=10 ? 3 : n%100>=11 ? 4 : 5;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "#. Comments\n"
-            + "#: core/logic/week_in_review_email_logic.py:49\n"
-            + "msgid \"repin \\\"{placeholder}\\\"\"\n"
-            + "msgid_plural \"repins \\\"{placeholder}\\\"\"\n"
-            + "msgstr[0] \"repins \\\"{placeholder}\\\"-ar-0\"\n"
-            + "msgstr[1] \"repins \\\"{placeholder}\\\"-ar-1\"\n"
-            + "msgstr[2] \"repins \\\"{placeholder}\\\"-ar-2\"\n"
-            + "msgstr[3] \"repins \\\"{placeholder}\\\"-ar-3\"\n"
-            + "msgstr[4] \"repins \\\"{placeholder}\\\"-ar-4\"\n"
-            + "msgstr[5] \"repins \\\"{placeholder}\\\"-ar-5\"\n";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ar-AR",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeMacStringsRemoveUntranslated() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repo, "ja-JP");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        """
-            /* comment 1 */
-            "key1" = "value1";
-
-            /* comment 2 */
-            "key2" = "value2";
-            """;
-
-    String expectedLocalizedAsset = "\n";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "Localizable.strings", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.debug("localized=\n{}\nEOL", localizedAsset);
-    assertEquals(expectedLocalizedAsset, localizedAsset);
-
-    String forImport =
-        """
-            /* comment 1 */
-            "key1" = "value1-jp";
-
-            /* comment 2 */
-            "key2" = "value2-jp";
-            """;
-
-    logger.debug("formimport=\n{}", forImport);
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.info("localized after import=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeMacStringsNamessNotEnclosedInDoubleQuotes() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent = "NSUsageDescription = \"Usage description:\";\n";
-    asset =
-        assetService.createAssetWithContent(
-            repo.getId(), "en.lproj/Localizable.strings", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    assertEquals(1, textUnitDTOs.size());
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeXtb() throws Exception {
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<!DOCTYPE translationbundle>\n"
-            + "<translationbundle lang=\"en-US\">\n"
-            + "	<translation id=\"0\" key=\"MSG_DIALOG_OK_\" source=\"lib/closure-library/closure/goog/ui/dialog.js\" desc=\"Standard caption for the dialog 'OK' button.\">OK</translation>\n"
-            + "     <translation id=\"1\" key=\"MSG_VIEWER_MENU\" source=\"src/js/box/dicom/viewer/toolbar.js\" desc=\"Tooltip text for the &quot;More&quot; menu.\">More</translation>\n"
-            + "     <translation id=\"2\" key=\"MSG_GONSTEAD_STEP\" source=\"src/js/box/dicom/viewer/gonsteaddialog.js\" desc=\"Instructions for the Gonstead method.\">Select the &lt;strong&gt;left Iliac crest&lt;/strong&gt;</translation>\n"
-            + "</translationbundle>";
-    String expectedContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<translationbundle lang=\"en-US\">\n"
-            + "	<translation desc=\"Standard caption for the dialog 'OK' button.\" id=\"0\" key=\"MSG_DIALOG_OK_\" source=\"lib/closure-library/closure/goog/ui/dialog.js\">OK</translation>\n"
-            + "     <translation desc=\"Tooltip text for the &quot;More&quot; menu.\" id=\"1\" key=\"MSG_VIEWER_MENU\" source=\"src/js/box/dicom/viewer/toolbar.js\">More</translation>\n"
-            + "     <translation desc=\"Instructions for the Gonstead method.\" id=\"2\" key=\"MSG_GONSTEAD_STEP\" source=\"src/js/box/dicom/viewer/gonsteaddialog.js\">Select the &lt;strong&gt;left Iliac crest&lt;/strong&gt;</translation>\n"
-            + "</translationbundle>";
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "xtb/messages-en-US.xtb", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    assertEquals(3, textUnitDTOs.size());
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("source=[{}]", textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedContent, localizedAsset);
-  }
-
-  @Ignore("Bug: does not output translationbundle opening tag")
-  @Test
-  public void testLocalizeXtbRemoveUntranslated() throws Exception {
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<!DOCTYPE translationbundle>\n"
-            + "<translationbundle lang=\"en-US\">\n"
-            + "	<translation id=\"0\" key=\"MSG_DIALOG_OK_\" source=\"lib/closure-library/closure/goog/ui/dialog.js\" desc=\"Standard caption for the dialog 'OK' button.\">OK</translation>\n"
-            + "     <translation id=\"1\" key=\"MSG_VIEWER_MENU\" source=\"src/js/box/dicom/viewer/toolbar.js\" desc=\"Tooltip text for the &quot;More&quot; menu.\">More</translation>\n"
-            + "     <translation id=\"2\" key=\"MSG_GONSTEAD_STEP\" source=\"src/js/box/dicom/viewer/gonsteaddialog.js\" desc=\"Instructions for the Gonstead method.\">Select the &lt;strong&gt;left Iliac crest&lt;/strong&gt;</translation>\n"
-            + "</translationbundle>";
-    String expectedContent =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" + "<translationbundle />\n";
-
-    asset =
-        assetService.createAssetWithContent(repo.getId(), "xtb/messages-en-US.xtb", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    // assertEquals(expectedContent, localizedAsset);
-
-    String forImport =
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-            + "<!DOCTYPE translationbundle>\n"
-            + "<translationbundle lang=\"ja-JP\">\n"
-            + "<translation id=\"1\" key=\"MSG_VIEWER_MENU\" source=\"src/js/box/dicom/viewer/toolbar.js\" desc=\"Tooltip text for the &quot;More&quot; menu.\">Plus</translation>\n"
-            + "</translationbundle>";
-
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            forImport,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.TRANSLATION_NEEDED,
-            null,
-            null)
-        .get();
-
-    localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "ja-JP",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.REMOVE_UNTRANSLATED,
-            null);
-    logger.debug("localized after import=\n{}", localizedAsset);
-    assertEquals(forImport, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeJsonWithComments() throws Exception {
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "{\n"
-            + "  // Greeting from Main UI 1\n"
-            + "  \"hello1\": \"Hello 1\",\n"
-            + "  \"hello2\": \"Hello 2\",\n"
-            + "  // Extra\n"
-            + "  // Greeting from Main UI 3\n"
-            + "  \"hello3\": \"Hello 3\"\n"
-            + "}";
-    String expectedContent = assetContent;
-
-    asset = assetService.createAssetWithContent(repo.getId(), "strings.json", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    assertEquals(3, textUnitDTOs.size());
-    assertEquals("Greeting from Main UI 1", textUnitDTOs.get(0).getComment());
-    assertNull(textUnitDTOs.get(1).getComment());
-    assertEquals("Greeting from Main UI 3", textUnitDTOs.get(2).getComment());
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("name=[{}], source=[{}]", textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeJson() throws Exception {
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    List<String> jsonFilterOptions =
-        Arrays.asList("useFullKeyPath=true", "extractAllPairs=false", "exceptions=.*/string");
-
-    String assetContent =
-        "{\n"
-            + "  \"this to ignore\": {\n"
-            + "    \"k1\": \"v1\"\n"
-            + "  },\n"
-            + "  \"hello_world\": {\n"
-            + "    \"string\": \"Hello World\",\n"
-            + "    \"note\": \"The start of every language book.\"\n"
-            + "  },\n"
-            + "  \"num_photos\": {\n"
-            + "    \"string\": \"You have {numPhotos, plural, =0 {no photos.} =1 {one photo.} other {# photos.}}\",\n"
-            + "    \"note\": \"A description that shows the number of photos a user has.\"\n"
-            + "  }\n"
-            + "}";
-    String expectedContent = assetContent;
-
-    asset = assetService.createAssetWithContent(repo.getId(), "strings.json", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(),
-            asset.getPath(),
-            assetContent,
-            false,
-            null,
-            null,
-            null,
-            null,
-            null,
-            jsonFilterOptions);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    assertEquals(2, textUnitDTOs.size());
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("name=[{}], source=[{}]", textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    assertEquals("hello_world/string", textUnitDTOs.get(0).getName());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            jsonFilterOptions,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeJsonRemoveKeySuffix() throws Exception {
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    List<String> jsonFilterOptions =
-        Arrays.asList(
-            "useFullKeyPath=true",
-            "extractAllPairs=false",
-            "exceptions=.*/string",
-            "removeKeySuffix=/string");
-
-    String assetContent =
-        "{\n"
-            + "  \"this to ignore\": {\n"
-            + "    \"k1\": \"v1\"\n"
-            + "  },\n"
-            + "  \"hello_world\": {\n"
-            + "    \"string\": \"Hello World\",\n"
-            + "    \"note\": \"The start of every language book.\"\n"
-            + "  },\n"
-            + "  \"num_photos\": {\n"
-            + "    \"string\": \"You have {numPhotos, plural, =0 {no photos.} =1 {one photo.} other {# photos.}}\",\n"
-            + "    \"note\": \"A description that shows the number of photos a user has.\"\n"
-            + "  }\n"
-            + "}";
-    String expectedContent = assetContent;
-
-    asset = assetService.createAssetWithContent(repo.getId(), "strings.json", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(),
-            asset.getPath(),
-            assetContent,
-            false,
-            null,
-            null,
-            null,
-            null,
-            null,
-            jsonFilterOptions);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    assertEquals(2, textUnitDTOs.size());
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug("name=[{}], source=[{}]", textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    assertEquals("hello_world", textUnitDTOs.get(0).getName());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            jsonFilterOptions,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeJsonHTMLCode() throws Exception {
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    List<String> jsonFilterOptions =
-        Arrays.asList(
-            "convertToHtmlCodes=true",
-            "codeFinderData=#v1\n"
-                + "count.i=3\n"
-                + "rule0=%(([-0+#]?)[-0+#]?)((\\d\\$)?)(([\\d\\*]*)(\\.[\\d\\*]*)?)[dioxXucsfeEgGpn]\n"
-                + "rule1=(\\\\r\\\\n)|\\\\a|\\\\b|\\\\f|\\\\n|\\\\r|\\\\t|\\\\v\n"
-                + "rule2=\\{\\d.*?\\}\n"
-                + "sample=%s, %d, {1}, \\n, \\r, \\t, etc.\n"
-                + "useAllRulesWhenTesting.b=false");
-
-    String assetContent = "{\"hello\" : \"Hello %s!\" }\n";
-
-    String expectedContent = assetContent;
-    asset = assetService.createAssetWithContent(repo.getId(), "strings.json", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(),
-            asset.getPath(),
-            assetContent,
-            false,
-            null,
-            null,
-            null,
-            null,
-            null,
-            jsonFilterOptions);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    assertEquals(1, textUnitDTOs.size());
-    assertEquals("Hello <br id='p1'/>!", textUnitDTOs.get(0).getSource());
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.info("name=[{}], source=[{}]", textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            jsonFilterOptions,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(expectedContent, localizedAsset);
-  }
-
   @Test
   public void testImportLocalizedAssetApproved()
       throws RepositoryNameAlreadyUsedException,
@@ -4462,123 +1674,6 @@ public class TMServiceTest extends ServiceTestBase {
   }
 
   @Test
-  public void testImportLocalizedAssetPoPluralSinglePlural() throws Exception {
-    repository = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale;
-    try {
-      repoLocale = repositoryService.addRepositoryLocale(repository, "ja-JP");
-    } catch (RepositoryLocaleCreationException e) {
-      throw new RuntimeException(e);
-    }
-
-    String assetContent =
-        "# SOME DESCRIPTIVE TITLE.\n"
-            + "# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER\n"
-            + "# This file is distributed under the same license as the PACKAGE package.\n"
-            + "# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.\n"
-            + "#\n"
-            + "#, fuzzy\n"
-            + "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-02-24 11:50-0800\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"Language: \\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=2; plural=(n != 1);\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + " \n"
-            + "   \n"
-            + "#. Test plural\n"
-            + "#: file.js:20\n"
-            + "msgctxt \"car\"\n"
-            + "msgid \"There is {number} car\"\n"
-            + "msgid_plural \"There are {number} cars\"\n"
-            + "msgstr[0] \"\"\n"
-            + "msgstr[1] \"\"\n"
-            + "\n"
-            + "#. Test plural okapi bug\n"
-            + "#: file.js:24\n"
-            + "msgctxt \"testpluralokapibug\"\n"
-            + "msgid \"test okapi bug\"\n"
-            + "msgstr \"\"\n"
-            + "";
-    asset = assetService.createAssetWithContent(repository.getId(), "messages.pot", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repository.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repository.getId(),
-            asset.getPath(),
-            assetContent,
-            false,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    String localizedAssetContent =
-        "# SOME DESCRIPTIVE TITLE.\n"
-            + "# Copyright (C) YEAR THE PACKAGE'S COPYRIGHT HOLDER\n"
-            + "# This file is distributed under the same license as the PACKAGE package.\n"
-            + "# FIRST AUTHOR <EMAIL@ADDRESS>, YEAR.\n"
-            + "#\n"
-            + "#, fuzzy\n"
-            + "msgid \"\"\n"
-            + "msgstr \"\"\n"
-            + "\"Project-Id-Version: PACKAGE VERSION\\n\"\n"
-            + "\"Report-Msgid-Bugs-To: \\n\"\n"
-            + "\"POT-Creation-Date: 2017-02-24 11:50-0800\\n\"\n"
-            + "\"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\\n\"\n"
-            + "\"Last-Translator: FULL NAME <EMAIL@ADDRESS>\\n\"\n"
-            + "\"Language-Team: LANGUAGE <LL@li.org>\\n\"\n"
-            + "\"Language: \\n\"\n"
-            + "\"MIME-Version: 1.0\\n\"\n"
-            + "\"Plural-Forms: nplurals=1; plural=0;\\n\"\n"
-            + "\"Content-Type: text/plain; charset=utf-8\\n\"\n"
-            + "\"Content-Transfer-Encoding: 8bit\\n\"\n"
-            + "  \n"
-            + "   \n"
-            + "#. Test plural\n"
-            + "#: file.js:20\n"
-            + "msgctxt \"car\"\n"
-            + "msgid \"There is {number} car\"\n"
-            + "msgid_plural \"There are {number} cars\"\n"
-            + "msgstr[0] \"There is {number} car\"\n"
-            + "\n"
-            + "#. Test plural okapi bug\n"
-            + "#: file.js:24\n"
-            + "msgctxt \"testpluralokapibug\"\n"
-            + "msgid \"test okapi bug\"\n"
-            + "msgstr \"jp test okapi bug\"\n"
-            + "\n"
-            + "";
-    tmService
-        .importLocalizedAssetAsync(
-            assetId,
-            localizedAssetContent,
-            repoLocale.getLocale().getId(),
-            StatusForEqualTarget.APPROVED,
-            null,
-            null)
-        .get();
-  }
-
-  @Test
   public void testImportLocalizedWithDuplicatedName() throws Exception {
     repository = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
     RepositoryLocale repoLocale;
@@ -4676,300 +1771,5 @@ public class TMServiceTest extends ServiceTestBase {
         pollableTaskService.getPollableTask(exportResult.getPollableTask().getId());
     tmXliff = tmXliffRepository.findByPollableTask(pollableTask);
     assertEquals(expected, tmXliff.getContent());
-  }
-
-  @Test
-  public void testLocalizeTSFile() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "namespace Translations {\n"
-            + "    export const en = {\n"
-            + "        // login comment\n"
-            + "        \"loginText\": \"Log In\",\n"
-            + "        // signup comment\n"
-            + "        \"signupText\": \"Sign up with `backquote`\",\n"
-            + "        \"quotedText\": \"Hello \\\"%s\\\"\",\n"
-            + "        \"noComment\": \"String with no comment\\nand newline\",\n"
-            + "        // template literals\n"
-            + "        \"templateText1\": `one line`,\n"
-            + "        \"templateText2\": `one line no comment`,\n"
-            + "        \"templateText3\": `one line \\`/special\\` character`,\n"
-            + "        // template multiline literals\n"
-            + "        \"templateMultilineText1\": `first line\nsecond line`,\n"
-            + "        // template multiline literals with escaped backquote\n"
-            + "        \"templateMultilineText2\": `special character\ncheck \\`/command\\` out`,\n\n"
-            + "    };\n"
-            + "}\n"
-            + "\n"
-            + "export default Translations;";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "translations.ts", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug(
-          "{}\n{}=[{}]", textUnitDTO.getComment(), textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    assertEquals(9, textUnitDTOs.size());
-    assertEquals("Sign up with `backquote`", textUnitDTOs.get(1).getSource());
-    assertEquals("Hello \"%s\"", textUnitDTOs.get(2).getSource());
-    assertEquals("String with no comment\nand newline", textUnitDTOs.get(3).getSource());
-    assertEquals("special character\ncheck `/command` out", textUnitDTOs.get(8).getSource());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeYAMLFile() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "activerecord:\n"
-            + "  errors:\n"
-            + "    template:\n"
-            + "      header:\n"
-            + "        list: [one, two, three]\n"
-            + "        map: {key: value, key2: value2}\n"
-            + "        one: \"Impossible d'enregistrer {{model}}: 1 erreur\"\n"
-            + "        other: \"Impossible d'enregistrer {{model}}: {{count}} erreurs.\"";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "translations.yaml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(), asset.getPath(), assetContent, false, null, null, null, null, null, null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.debug(
-          "{}\n{}=[{}]", textUnitDTO.getComment(), textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    assertEquals(7, textUnitDTOs.size());
-    assertEquals("one", textUnitDTOs.get(0).getSource());
-    assertEquals("two", textUnitDTOs.get(1).getSource());
-    assertEquals("three", textUnitDTOs.get(2).getSource());
-    assertEquals("value", textUnitDTOs.get(3).getSource());
-    assertEquals("value2", textUnitDTOs.get(4).getSource());
-    assertEquals("Impossible d'enregistrer {{model}}: 1 erreur", textUnitDTOs.get(5).getSource());
-    assertEquals(
-        "Impossible d'enregistrer {{model}}: {{count}} erreurs.", textUnitDTOs.get(6).getSource());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeYAMLFileWithFilterOptions() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    List<String> filterOptions =
-        Arrays.asList(
-            "extractAllPairs=false", "exceptions=one|activerecord/errors/template/header/other");
-
-    String assetContent =
-        "activerecord:\n"
-            + "  errors:\n"
-            + "    template:\n"
-            + "      header:\n"
-            + "        list: [one, two, three]\n"
-            + "        map: {key: value, key2: value2}\n"
-            + "        one: \"Impossible d'enregistrer {{model}}: 1 erreur\"\n"
-            + "        other: \"Impossible d'enregistrer {{model}}: {{count}} erreurs.\"";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "translations.yaml", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(),
-            asset.getPath(),
-            assetContent,
-            false,
-            null,
-            null,
-            null,
-            null,
-            null,
-            filterOptions);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.info(
-          "{}\n{}=[{}]", textUnitDTO.getComment(), textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    assertEquals(2, textUnitDTOs.size());
-    assertEquals("Impossible d'enregistrer {{model}}: 1 erreur", textUnitDTOs.get(0).getSource());
-    assertEquals(
-        "Impossible d'enregistrer {{model}}: {{count}} erreurs.", textUnitDTOs.get(1).getSource());
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            null,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.debug("localized=\n{}", localizedAsset);
-    assertEquals(assetContent, localizedAsset);
-  }
-
-  @Test
-  public void testLocalizeHtmlFilter() throws Exception {
-
-    Repository repo = repositoryService.createRepository(testIdWatcher.getEntityName("repository"));
-    RepositoryLocale repoLocale = repositoryService.addRepositoryLocale(repo, "en-GB");
-
-    String assetContent =
-        "<!DOCTYPE html>\n"
-            + "<html>\n"
-            + "<head>\n"
-            + "    <title>My Title</title>\n"
-            + "    <meta name=\"description\" content=\"My description\"/>\n"
-            + "    <meta name=\"author\" content=\"My author\"/>\n"
-            + "    <meta name=\"keywords\" content=\"My keywords\"/>\n"
-            + "    <link rel=\"stylesheet\" href=\"./stylesheet.css\" type=\"text/css\"/>\n"
-            + "    <style>.body {\n"
-            + "        width: auto;\n"
-            + "    }</style>\n"
-            + "</head>\n"
-            + "<body>\n"
-            + "<p>thi is the first paragraph</p>\n"
-            + "<p>this is the second paragraph. With an <img src=\"someimage.jpg\"> inside text</p>\n"
-            + "<ul>\n"
-            + "    <li>item1</li>\n"
-            + "    <li>item2</li>\n"
-            + "</ul>\n"
-            + "</body>\n"
-            + "</html>";
-
-    asset = assetService.createAssetWithContent(repo.getId(), "demo.html", assetContent);
-    asset = assetRepository.findById(asset.getId()).orElse(null);
-    assetId = asset.getId();
-    tmId = repo.getTm().getId();
-
-    PollableFuture<Asset> assetResult =
-        assetService.addOrUpdateAssetAndProcessIfNeeded(
-            repo.getId(),
-            asset.getPath(),
-            assetContent,
-            false,
-            null,
-            null,
-            null,
-            null,
-            FilterConfigIdOverride.HTML_ALPHA,
-            null);
-    try {
-      pollableTaskService.waitForPollableTask(assetResult.getPollableTask().getId());
-    } catch (PollableTaskException | InterruptedException e) {
-      throw new RuntimeException(e);
-    }
-    assetResult.get();
-
-    TextUnitSearcherParameters textUnitSearcherParameters = new TextUnitSearcherParameters();
-    textUnitSearcherParameters.setRepositoryIds(repo.getId());
-    textUnitSearcherParameters.setStatusFilter(StatusFilter.FOR_TRANSLATION);
-    List<TextUnitDTO> textUnitDTOs = textUnitSearcher.search(textUnitSearcherParameters);
-    for (TextUnitDTO textUnitDTO : textUnitDTOs) {
-      logger.info(
-          "{}\n{}=[{}]", textUnitDTO.getComment(), textUnitDTO.getName(), textUnitDTO.getSource());
-    }
-
-    String localizedAsset =
-        tmService.generateLocalized(
-            asset,
-            assetContent,
-            repoLocale,
-            "en-GB",
-            FilterConfigIdOverride.HTML_ALPHA,
-            null,
-            Status.ALL,
-            InheritanceMode.USE_PARENT,
-            null);
-    logger.info("localized=\n{}", localizedAsset);
-
-    // Okapi adds meta tag in that case? this could be a problem or not, just putting a note here
-    // for now. Did not
-    // see that happen in CLI tests
-    localizedAsset =
-        localizedAsset.replace(
-            "<meta http-equiv=\"Content-Type\" content=\"text/html; charset=utf-8\">", "");
-    assertEquals(assetContent, localizedAsset);
   }
 }
